requirements.txt
flask
python-dotenv
Flask-SQLAlchemy
psycopg2
flask-marshmallow
marshmallow-sqlalchemy
passlib[bcrypt]
python-jose[cryptography]